cluster/health: move node-local definitions
Creates node/types.[cc,h] and moves node-local health structs there.
