+++ src/v/cluster/CMakeLists.txt
@@ -85,6 +85,7 @@ v_cc_library(
     health_monitor_backend.cc
     health_monitor_frontend.cc
     metrics_reporter.cc
+    node/types.cc
   DEPS
     Seastar::seastar
     controller_rpc
